move add form into separate file for general use
