+++ frontend/awx/resources/inventories/InventoryGroups/GroupPageAdd.cy.tsx
